Move AccountSetupTopAppBar to common and rename to AccountTopAppBar
